Fix: story page in layout + keep intr0.tsx naming
