+++ content/chapters/chapter-1/index.tsx
@@ -16,17 +16,16 @@ export const metadata = {
   slug: 'chapter-1',
   title: 'chapter_one.title',
   image: '/assets/images/chapter-1-cover.jpg',
-  intros: ['genesis-1', 'genesis-2'],
+  intros: ['intro-1', 'intro-2'],
   lessons: [
     'genesis-1',
     'genesis-2',
     'genesis-3',
     'genesis-4',
-    'genesis-5',
     'transacting-1',
     'transacting-2',
     'transacting-3',
   ],
-  outros: ['outro-1'],
+  outros: ['outro-1, outro-2'],
   challenges: ['genesis-1', 'transacting-1'],
 }
