+++ content/lessons/chapter-1/index.ts
@@ -1,9 +1,9 @@
+import * as Intro1 from './intro-1'
+import * as Intro2 from './intro-2'
 import * as Genesis1 from './genesis-1'
 import * as Genesis2 from './genesis-2'
 import * as Genesis3 from './genesis-3'
 import * as Genesis4 from './genesis-4'
-import * as Genesis5 from './genesis-5'
-import * as Genesis6 from './genesis-6'
 import * as Transacting1 from './transacting-1'
 import * as Transacting2 from './transacting-2'
 import * as Transacting3 from './transacting-3'
@@ -11,12 +11,12 @@ import * as Outro1 from './outro-1'
 import * as Outro2 from './outro-2'
 
 const Chapter1 = {
+  'intro-1': Intro1,
+  'intro-2': Intro2,
   'genesis-1': Genesis1,
   'genesis-2': Genesis2,
   'genesis-3': Genesis3,
   'genesis-4': Genesis4,
-  'genesis-5': Genesis5,
-  'genesis-6': Genesis6,
   'transacting-1': Transacting1,
   'transacting-2': Transacting2,
   'transacting-3': Transacting3,
