+++ content/lessons/chapter-1/intro-2.tsx
@@ -1,0 +1,35 @@
+'use client'
+
+import { useLocalizedRoutes, useTranslations } from 'hooks'
+import { Introduction, Text } from 'ui'
+
+export const metadata = {
+  title: 'chapter_one.intro_two.title',
+  image: '/assets/images/chapter-1-holocat.jpg',
+  theme: 'bg-[#32303D]',
+}
+
+export default function Intro2({ lang }) {
+  const routes = useLocalizedRoutes()
+  const t = useTranslations(lang)
+
+  return (
+    <Introduction lang={lang}>
+      <Text className="text-lg md:text-xl">
+        {t('chapter_one.intro_two.paragraph_one')}
+      </Text>
+      <Text className="mt-4 text-lg md:text-xl">
+        {t('chapter_one.intro_two.paragraph_two')}
+      </Text>
+      <Text className="mt-4 text-lg md:text-xl">
+        {t('chapter_one.intro_two.paragraph_three')}
+      </Text>
+      <Text className="mt-4 text-lg md:text-xl">
+        {t('chapter_one.intro_two.paragraph_four')}
+      </Text>
+      <Text className="mt-4 text-lg md:text-xl">
+        {t('chapter_one.intro_two.paragraph_five')}
+      </Text>
+    </Introduction>
+  )
+}
