+++ content/lessons/chapter-2/index.ts
@@ -1,11 +1,11 @@
-import * as Hashing1 from './hashing-1'
-import * as Hashing2 from './hashing-2'
+import * as Intro1 from './intro-1'
+import * as Intro2 from './intro-2'
 import * as Scripting1 from './scripting-1'
 import * as Scripting2 from './scripting-2'
 
 const Chapter2 = {
-  'hashing-1': Hashing1,
-  'hashing-2': Hashing2,
+  'intro-1': Intro1,
+  'intro-2': Intro2,
   'scripting-1': Scripting1,
   'scripting-2': Scripting2,
 }
